+++ src/main/java/com/efub/series/domain/content/controller/ComicController.java
@@ -1,8 +1,8 @@
-package com.efub.series.domain.comic.controller;
+package com.efub.series.domain.content.controller;
 
-import com.efub.series.domain.comic.domain.Content;
-import com.efub.series.domain.comic.dto.ComicListResDto;
-import com.efub.series.domain.comic.service.ComicService;
+import com.efub.series.domain.content.domain.Content;
+import com.efub.series.domain.content.dto.ComicListResDto;
+import com.efub.series.domain.content.service.ContentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -14,19 +14,19 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ComicController {
 
-    private final ComicService comicService;
+    private final ContentService contentService;
 
     @GetMapping
     @ResponseStatus(value = HttpStatus.OK)
-    public ComicListResDto  getComicList(){
-        List<Content> comicList = comicService.findTop5ByType("comic");
+    public ComicListResDto getContentList(){
+        List<Content> comicList = contentService.findTop5ByType("comic");
         return ComicListResDto.of(comicList);
     }
 
     @GetMapping("/freeList")
     @ResponseStatus(value = HttpStatus.OK)
     public ComicListResDto getComicFreeList(@RequestParam String freeType){
-        List<Content> comicFreeList = comicService.findTop5ByTypeAndFreeType("comic", freeType);
+        List<Content> comicFreeList = contentService.findTop5ByTypeAndFreeType("comic", freeType);
         return ComicListResDto.of(comicFreeList);
     }
 }
