+++ src/main/java/com/efub/series/domain/content/controller/NovelController.java
@@ -1,8 +1,8 @@
-package com.efub.series.domain.novel.controller;
+package com.efub.series.domain.content.controller;
 
-import com.efub.series.domain.novel.domain.Content;
-import com.efub.series.domain.novel.dto.NovelListResDto;
-import com.efub.series.domain.novel.service.NovelService;
+import com.efub.series.domain.content.domain.Content;
+import com.efub.series.domain.content.dto.NovelListResDto;
+import com.efub.series.domain.content.service.ContentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -14,19 +14,19 @@ import java.util.List;
 @RequiredArgsConstructor
 public class NovelController {
 
-    private final NovelService novelService;
+    private final ContentService contentService;
 
     @GetMapping
     @ResponseStatus(value = HttpStatus.OK)
     public NovelListResDto  getNovelList(){
-        List<Content> novelList = novelService.findTop5ByType("novel");
+        List<Content> novelList = contentService.findTop5ByType("novel");
         return NovelListResDto.of(novelList);
     }
 
     @GetMapping("/freeList")
     @ResponseStatus(value = HttpStatus.OK)
     public NovelListResDto getNovelFreeList(@RequestParam String freeType){
-        List<Content> novelFreeList = novelService.findTop5ByTypeAndFreeType("novel", freeType);
+        List<Content> novelFreeList = contentService.findTop5ByTypeAndFreeType("novel", freeType);
         return NovelListResDto.of(novelFreeList);
     }
 }
