+++ src/main/java/com/efub/series/domain/content/repository/ContentRepository.java
@@ -1,11 +1,11 @@
-package com.efub.series.domain.novel.repository;
+package com.efub.series.domain.content.repository;
 
-import com.efub.series.domain.novel.domain.Content;
+import com.efub.series.domain.content.domain.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface NovelRepository extends JpaRepository<Content, Long> {
+public interface ContentRepository extends JpaRepository<Content, Long> {
     List<Content> findTop5ByType(String type);
     List<Content> findTop5ByTypeAndFreeType(String type, String freeType);
 }
